Finished basic device chapter 9 support.
git-svn-id: http://lufa-lib.googlecode.com/svn@16 d5102386-fcda-11dd-9fdb-3debd5008f28

+++ MyUSB/Drivers/USB/LowLevel/LowLevel.h
@@ -98,7 +98,7 @@
 		
 		static inline void USB_Ignore_Word(void)
 		{
-			volatile uint16_t Dummy;
+			volatile uint8_t Dummy;
 			
 			Dummy = UEDATX;
 			Dummy = UEDATX;
